pyproject.toml
[tool.poetry]
name = "typhoon"
version = "0.1.0"
description = "On-demand, asset specific, neural network trianing microservice"
authors = ["otgon <otgon.petru@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.13,<=3.9.18"
pandas = "^2.2.1"
numpy = "^1.26.4"
fastapi = "^0.110.1"
confluent-kafka = "^2.3.0"
python-dotenv = "^1.0.1"
mysqlclient = "^2.2.4"
tortoise-orm = "^0.20.0"
matplotlib = "^3.8.3"
tensorflow-io-gcs-filesystem = "0.27.0"
tensorflow = "~2.10"
aerich = "^0.7.2"
aiomysql = "^0.2.0"
yfinance = "^0.2.37"
aiofiles = "^23.2.1"
datetime = "^5.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
uvicorn = "^0.29.0"
pydantic = "^2.7.0"
python-multipart = "^0.0.9"
pytest = "^8.1.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"
asgi-lifespan = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
